+++ R/MsBackendCompDb.R
@@ -21,6 +21,8 @@
 #'     `"precursor_mz"`, `"precursor_intensity"`, `"precursor_charge"` are
 #'     mapped to the core `Spectra` variables `msLevel`, `precursorMz`,
 #'     `precursorIntensity` and `precursorCharge`, respectively.
+#'     For `peaksData`: `character` with the names of the peaks columns to
+#'     return. Use `peaksVariables` for supported values.
 #'
 #' @param drop For `[`: not considered.
 #'
@@ -60,6 +62,9 @@
 #'   length equal to the number of spectra and each element being a `matrix`
 #'   with columns `"mz"` and `"intensity"` with the spectra's m/z and intensity
 #'   values.
+#'
+#' - `peaksVariables`: lists the available peaks variables in the backend
+#'   (database). These can be used for parameter `columns` of `peaksData`.
 #'
 #' - `intensity<-`: not supported.
 #'
@@ -185,10 +190,20 @@ setMethod("show", "MsBackendCompDb", function(object) {
 #'
 #' @exportMethod peaksData
 #'
+#' @importMethodsFrom Spectra peaksVariables
+#'
 #' @rdname MsBackendCompDb
-setMethod("peaksData", "MsBackendCompDb", function(object) {
-    .peaks_data(object)
+setMethod(
+    "peaksData", "MsBackendCompDb",
+    function(object, columns = c("mz", "intensity")) {
+        .peaks_data(object, columns = columns)
 })
+
+#' @exportMethod peaksVariables
+#'
+#' @rdname MsBackendCompDb
+setMethod("peaksVariables", "MsBackendCompDb",
+          function(object) .available_peaks_variables(object))
 
 #' @importMethodsFrom ProtGenerics dataStorage
 #'